Add static ip to service
A static ip is needed for the okapi gateway so that we can have a reliable address to connect to. We can also point a URL and all the modules at this static ip. During the Kubernetes Service creation it will use this static ip address that is reserved in Google Container Engine.