Create an ingress object and ingress controller using nginx
To use SSL/TLS and get an HTTPS connection we needed to create a L7 LoadBalancer using nginx. Initially creating a LoadBalancer on the Okapi service only created an L4 LoadBalancer which was not able to terminate SSL/TLS connections. An ingress object is used in conjuction with an ingress controller to watch for incoming request and send those request to the Okapi backend service while adding the appropriate TLS certs needed.