Change service to NodePort
Initially we had Google Container Engine generate an external loadBalancer (ELB) using a Kubernetes Service that gave us an external ip address that we used to access the okapi gatewway running in the cluster behind this service. However as the need for SSL/TLS came to the forefront we could not rely on the ELB that Google Container Engine generated with the Service because it was an L4 loadBalancer and L4 cannot terminate SSL/TLS so we need a L7. Changing the port allows this to be configured to work with the nginx ingress that is setup in the cluster.